Update gen_wg_config.sh, add swithc function and cleaned code up a bit

+++ gen_wg_config.sh
@@ -1,26 +1,26 @@
 #!/bin/sh
 set -e
 
-read_config1() {
+read_config() {
     config_folder=$(dirname $(readlink -f $0))
 
-    config_file=${config_folder}/"config.json"
-    conf_json=$(cat $config_file)
-   
-    username="$(echo $conf_json | jq '.username')"
-    username=$(eval echo $username)
-
-    password="$(echo $conf_json | jq '.password')"
-    password=$(eval echo $password)
+    username=$(eval echo $(jq '.username' ${config_folder}/config.json))
+    password=$(eval echo $(jq '.password' ${config_folder}/config.json))
 
     wg_keys="${config_folder}/wg.json"
     token_file="${config_folder}/token.json"
+    token_expires="${config_folder}/token_expires.json"
 
     baseurl_1="https://api.surfshark.com"
     baseurl_2="https://ux.surfshark.com"
     baseurl_3="https://api.uymgg1.com"
     baseurl_4="https://ux.uymgg1.com"
     urlcount=4
+
+    generic_servers_file="${config_folder}/generic_servers.json"
+    static_servers_file="${config_folder}/static_servers.json"
+    obfuscated_servers_file="${config_folder}/obfuscated_servers.json"
+    double_servers_file="${config_folder}/double_servers.json"    
 
     force_register=0
     register=1
@@ -29,8 +29,6 @@ read_config1() {
     wireguard_down=0
     generate_servers=0
     switch_conf=0
-
-    unset conf_json
 }
 
 parse_arg() {
@@ -40,14 +38,14 @@ parse_arg() {
             f)  force_register=1    ;;
             g)  generate_conf=0     ;;
             r)  generate_servers=1  ;;
-#            s)  switch_conf=1       ;;
+            s)  switch_conf=1       ;;
             ?|h)
             echo "Usage: $(basename $0) [-f]"
             echo "  -f force register, ignore checking"
             echo "  -g skip generating server conf files"
             echo "  -r regenerate the server conf files"
-#            echo "  -s switch from one surfshark wireguard conf to another"
-            echo "  -C clear keys and profile files before generating new ones"
+            echo "  -s switch from one surfshark wireguard server to another"
+            echo "  -C clear settings, keys and profile files before generating new ones"
             exit 1                  ;;
         esac
     done
@@ -57,38 +55,33 @@ parse_arg() {
 wg_login() {
 #add in renewal option
 #/v1/auth/renew
-    if [ -f "$token_file" ]; then
-        curl_res=$(cat $token_file)
-    else
+    http_status=0
+    basen=0
+    until [ $http_status -eq 200 ]; do
         tmpfile=$(mktemp /tmp/wg-curl-res.XXXXXX)
-        http_status=0
-        basen=0
-        until [ $http_status -eq 200 ]; do
-            let basen=$basen+1; baseurl=baseurl_$basen
-            if [ $basen -gt $urlcount ]; then
-                echo "Unable to login, check your credentials."
-                rm $tmpfile
-                exit 2
-            fi
-            url=$(eval echo \${$baseurl})/v1/auth/login
-            data="{\"username\":\"$username\", \"password\":\"$password\"}"
-            token=$(eval echo $token)
-            http_status=$(curl -o $tmpfile -s -w "%{http_code}" -d "$data" -H 'Content-Type: application/json' -X POST $url)
-            echo "Login "$url $http_status
-        done
+        let basen=$basen+1; baseurl=baseurl_$basen
+        if [ $basen -gt $urlcount ]; then
+            echo "Unable to login, check your credentials."
+            rm $tmpfile
+            exit 2
+        fi
+        url=$(eval echo \${$baseurl})/v1/auth/login
+        data="{\"username\":\"$username\", \"password\":\"$password\"}"
+        token=$(eval echo $token)
+        http_status=$(curl -o $tmpfile -s -w "%{http_code}" -d "$data" -H 'Content-Type: application/json' -X POST $url)
+        echo "Login "$url $http_status
         cp $tmpfile $token_file
         rm $tmpfile
-    fi
-    token=$(echo $curl_res | jq '.token')
-    renewToken=$(echo $curl_res | jq '.renewToken')
+    done
+    token=$(eval echo $(jq '.token' $token_file))
+    renewToken=$(eval echo $(jq '.renewToken' $token_file))
 }
 
 wg_gen_keys() {
     if [ -f "$wg_keys" ]; then
         echo "using existent wg keys"
-        wg_pub=$(cat $wg_keys | jq '.pub')
-        wg_prv=$(cat $wg_keys | jq '.prv')
-        wg_prv=$(eval echo $wg_prv)
+        wg_prv=$(eval echo $(jq '.prv' $wg_keys))
+        wg_pub=$(eval echo $(jq '.pub' $wg_keys))
     else 
         echo "generating new keys"
         wg_prv=$(wg genkey)
@@ -97,27 +90,16 @@ wg_gen_keys() {
     fi
 }
 
-read_config2() {
-    curl_res=$(cat $token_file)
-    token=$(echo $curl_res | jq '.token')
-    renewToken=$(echo $curl_res | jq '.renewToken')
-
-    generic_servers_file="${config_folder}/generic_servers.json"
-    static_servers_file="${config_folder}/static_servers.json"
-    obfuscated_servers_file="${config_folder}/obfuscated_servers.json"
-    double_servers_file="${config_folder}/double_servers.json"
-}
-
 wg_reg_pubkey() {
     curl_reg=401
     basen=1
-    error_count=0
+    error_count_et=0
+    error_count_nt=0
     while [ -z "${curl_reg##*401*}" ]; do
         baseurl=baseurl_$basen
         if [ $basen -gt $urlcount ]; then
-            echo "Token was not recognised, or Public Key was rejected please try again."
+            echo "Token was not recognised/expired or the Public Key was rejected please try again."
             echo "If it fails repeatedly check your credentials and that a token exists."
-
             exit 2
         fi
         url=$(eval echo \${$baseurl})/v1/account/users/public-keys
@@ -126,20 +108,19 @@ wg_reg_pubkey() {
         curl_reg=$(eval curl -H \"Authorization: Bearer $token\" -H \"Content-Type: application/json\" -d \'$data\' -X POST $url)
         echo "Registration "$url $curl_reg
         let basen=$basen+2
-        if [ -z "${curl_reg##*Expired*}" ]; then
+        if [ -z "${curl_reg##*Expired*}" ] && [ $error_count_et -eq 0 ]; then
             rm -f ${config_folder}/token.json ${config_folder}/wg.json  # temp solution
             wg_login                                                    # until renewal
             wg_gen_keys                                                 # is sorted
-            basen=1                                              #
-            continue                                                       #
-        elif [ -z "${curl_reg##*Token not found*}" ] && [ $error_count -eq 0 ]; then
-            curl_res=$(cat $token_file)
-            token=$(echo $curl_res | jq '.token')
-            renewToken=$(echo $curl_res | jq '.renewToken')
-            error_count=1
+            error_count_et=1                                            #
+            basen=1                                                     #
+        elif [ -z "${curl_reg##*Token not found*}" ] && [ $error_count_nt -eq 0 ]; then
+            token=$(eval echo $(jq '.token' $token_file))
+            renewToken=$(eval echo $(jq '.renewToken' $token_file))
+            error_count_nt=1
             basen=1
-        elif [ -z "${curl_reg##*Token not found*}" ] && [ $error_count -eq 1 ]; then
-            echo "Token was not recognised, or Public Key was rejected please try again."
+        elif [ -z "${curl_reg##*Token not found*}" ] && [ $error_count_nt -eq 1 ]; then
+            echo "Token was not recognised."
             echo "If it fails repeatedly check your credentials and that a token exists."
             exit 2
         fi
@@ -164,11 +145,10 @@ wg_check_pubkey() {
         http_status=$(eval curl -o $tmpfile -s -w "%{http_code}" -H \"Authorization: Bearer $token\" -H \"Content-Type: application/json\" -d \'$data\' -X POST $url)
         echo "Validation "$url $http_status
     done
-    curl_val=$(cat $tmpfile)
-    expire_date=$(echo $curl_val | jq '.expiresAt')
-    expire_date=$(eval echo $expire_date)
+    expire_date=$(eval echo $(jq '.expiresAt' $tmpfile))
     now=$(date -Iseconds --utc)
     if [ "${now}" '<' "${expire_date}" ];then
+        cp -f $tmpfile $token_expires
         register=0
         echo "TODAYS DATE"              # Display Run Date
         echo ""${now}""                 # and Time
@@ -227,8 +207,6 @@ gen_client_confs() {
             srv_pub="$(echo $row | jq '.[3]')"
             srv_pub=$(eval echo $srv_pub)
 
-#            echo "generating file for $srv_host"
-            
             file_name=${srv_host%$postf}
             file_name=${file_name/'-'/'-'$(printf %03d $srv_load)'-'}
             srv_tags=${srv_tags/'physical'/}
@@ -241,17 +219,7 @@ gen_client_confs() {
             else
 				file_name=${server}-${file_name}-${srv_tags}
 			fi
-
-			srv_conf_file=${config_folder}/conf/${file_name}.conf
-
-            srv_conf="[Peer]\ndescription=${srv_host%$postf}\npublic_key=$srv_pub\nendpoint_host=$srv_host"
-
-            if [ -f "$srv_conf_file" ]; then
-                echo -e "$srv_conf" > $srv_conf_file
-            else
-                echo -e "$srv_conf" >> $srv_conf_file
-            fi
-
+            "{\"description\":\"${srv_host%$postf}\",\"public_key\":\"$srv_pub\",\"endpoint_host\":\"$srv_host\"}" > ${config_folder}/conf/${file_name}.conf
         done
         file_removal="$server""_servers_file"
         file_removal=$(eval echo \${$file_removal})
@@ -259,32 +227,35 @@ gen_client_confs() {
     done
 }
 
-surfshark_up() {
+surfshark_1st_up() {
+    # allow fall back to wan if vpn off, can be overridden by kill switch
+    network_wan=$(grep -w -e network.wan.metric=\"1024\" /etc/config/network)
+    if [ ! $(IFS=.;set -f -- $network_test1; echo $#) -gt 0 ]; then
+        uci set network.wan.metric="1024"
+    fi
+
     network_test1=$(grep -w -e surfshark /etc/config/network)
     network_test2=$(grep -w -e wireguard /etc/config/network)
     if [ ! $(set -f -- $network_test1; echo $#) -gt 0 ] || [ ! $(set -f -- $network_test2; echo $#) -gt 0 ]; then
-        wg_prv=$(cat $wg_keys | jq '.prv')
-        wg_prv=$(eval echo $wg_prv)
-        uci add network device
-        uci set network.@device[-1]=device
-        uci set network.@device[-1].name='surfshark'
-        uci set network.@device[-1].ipv6='0'
-        uci set network.@device[-1].promisc='1'
-        uci set network.@device[-1].acceptlocal='1'
+        wg_prv=$(eval echo $(jq '.prv' $wg_keys))
+        uci set network.sswg=device
+        uci set network.sswg.name='surfshark'
+        uci set network.sswg.ipv6='0'
+        uci set network.sswg.promisc='1'
+        uci set network.sswg.acceptlocal='1'
         uci set network.surfshark=interface
         uci set network.surfshark.proto='wireguard'
         uci set network.surfshark.private_key=${wg_prv}
         uci set network.surfshark.listen_port='51820'
         uci set network.surfshark.addresses='10.14.0.2/8'
-        uci add network wireguard_surfshark
-        uci set network.@wireguard_surfshark[-1]=wireguard_surfshark
-        uci set network.@wireguard_surfshark[-1].description='wgs'
-        uci set network.@wireguard_surfshark[-1].public_key='o07k/2dsaQkLLSR0dCI/FUd3FLik/F/HBBcOGUkNQGo='
-        uci set network.@wireguard_surfshark[-1].allowed_ips='172.16.0.36/32'
-        uci set network.@wireguard_surfshark[-1].route_allowed_ips='1'
-        uci set network.@wireguard_surfshark[-1].endpoint_host='wgs.prod.surfshark.com'
-        uci set network.@wireguard_surfshark[-1].endpoint_port='51820'
-        uci set network.@wireguard_surfshark[-1].persistent_keepalive='25'
+        uci set network.wgs_surfshark=wireguard_surfshark
+        uci set network.wgs_surfshark.description='wgs'
+        uci set network.wgs_surfshark.public_key='o07k/2dsaQkLLSR0dCI/FUd3FLik/F/HBBcOGUkNQGo='
+        uci set network.wgs_surfshark.allowed_ips='172.16.0.36/32'
+        uci set network.wgs_surfshark.route_allowed_ips='1'
+        uci set network.wgs_surfshark.endpoint_host='wgs.prod.surfshark.com'
+        uci set network.wgs_surfshark.endpoint_port='51820'
+        uci set network.wgs_surfshark.persistent_keepalive='25'
         uci commit network
         /etc/init.d/network restart
     else
@@ -294,19 +265,17 @@ surfshark_up() {
     firewall_test1=$(grep -w -e surfsharkwg /etc/config/firewall)
     firewall_test2=$(grep -w -e surfshark /etc/config/firewall)
     if [ ! $(set -f -- $firewall_test1; echo $#) -gt 0 ] || [ ! $(set -f -- $firewall_test2; echo $#) -gt 0 ]; then
-        uci add firewall zone
-        uci set firewall.@zone[-1]=zone
-        uci set firewall.@zone[-1].name='surfsharkwg'
-        uci set firewall.@zone[-1].input='REJECT'
-        uci set firewall.@zone[-1].output='ACCEPT'
-        uci set firewall.@zone[-1].forward='REJECT'
-        uci set firewall.@zone[-1].masq='1'
-        uci set firewall.@zone[-1].mtu_fix='1'
-        uci set firewall.@zone[-1].network='surfshark'
-        uci add firewall forwarding
-        uci set firewall.@forwarding[-1]=forwarding
-        uci set firewall.@forwarding[-1].src='lan'
-        uci set firewall.@forwarding[-1].dest='surfsharkwg'
+        uci set firewall.sswg_zone=zone
+        uci set firewall.sswg_zone.name='surfsharkwg'
+        uci set firewall.sswg_zone.input='REJECT'
+        uci set firewall.sswg_zone.output='ACCEPT'
+        uci set firewall.sswg_zone.forward='REJECT'
+        uci set firewall.sswg_zone.masq='1'
+        uci set firewall.sswg_zone.mtu_fix='1'
+        uci set firewall.sswg_zone.network='surfshark'
+        uci set firewall.sswg_forwarding=forwarding
+        uci set firewall.sswg_forwarding.src='lan'
+        uci set firewall.sswg_forwarding.dest='surfsharkwg'
         uci commit firewall
         /etc/init.d/firewall restart
     else
@@ -316,40 +285,70 @@ surfshark_up() {
     echo "$(ls -xA ${config_folder}/conf/)"
     read -p "Please enter your choice of server: " selection
     if [ -f ${config_folder}/conf/${selection} ]; then
-        peer_desc=$(awk -F '=' '$1 ~ /^d/ {print $2}' ${config_folder}/conf/${selection})
-        peer_key=$(awk -F '=' '$1 ~ /^p/ {print $2}' ${config_folder}/conf/${selection})
-        peer_host=$(awk -F '=' '$1 ~ /^e/ {print $2}' ${config_folder}/conf/${selection})
-        uci add network wireguard_surfshark
-        uci set network.@wireguard_surfshark[-1]=wireguard_surfshark
-        uci set network.@wireguard_surfshark[-1].description=${peer_desc}
-        uci set network.@wireguard_surfshark[-1].public_key=${peer_key}=
-        uci set network.@wireguard_surfshark[-1].allowed_ips='0.0.0.0/0'
-        uci set network.@wireguard_surfshark[-1].route_allowed_ips='1'
-        uci set network.@wireguard_surfshark[-1].endpoint_host=${peer_host}
-        uci set network.@wireguard_surfshark[-1].endpoint_port='51820'
-        uci set network.@wireguard_surfshark[-1].persistent_keepalive='25'
+        peer_desc=$(eval echo $(jq '.description' ${config_folder}/conf/${selection}))
+        peer_key=$(eval echo $(jq '.public_key' ${config_folder}/conf/${selection}))
+        peer_host=$(eval echo $(jq '.endpoint_host' ${config_folder}/conf/${selection}))
+        uci set network.peer_surfshark=wireguard_surfshark
+        uci set network.peer_surfshark.description=${peer_desc}
+        uci set network.peer_surfshark.public_key=${peer_key}
+        uci set network.peer_surfshark.allowed_ips='0.0.0.0/0'
+        uci set network.peer_surfshark.route_allowed_ips='1'
+        uci set network.peer_surfshark.endpoint_host=${peer_host}
+        uci set network.peer_surfshark.endpoint_port='51820'
+        uci set network.peer_surfshark.persistent_keepalive='25'
         uci commit network
         /etc/init.d/network restart
-        echo "${config_folder}/conf/${selection}" >> ${config_folder}/surfshark
+        cp "${config_folder}/conf/${selection}" ${config_folder}/surfshark
     else
         echo "server conf not recognised"
         surfshark_up
     fi
 }
 
-reset_surfshark() {
-#    if [ -e ${config_folder}/surfshark ]; then
-#        rm ${config_folder}/surfshark
-#    fi
-    if [ -e ${config_folder}/wg.json ]; then
-        echo "Clearing old settings ..."
-        rm -fr ${config_folder}/conf ${config_folder}/*servers.json ${config_folder}/token.json ${config_folder}/wg.json
+surfshark_switch() {
+    echo "Current surfshark wireguard server is: "$(eval echo $(jq '.description' ${config_folder}/surfshark))
+    echo "$(ls -xA ${config_folder}/conf/)"
+    read -p "Please enter your choice of server: " selection
+    if [ -f ${config_folder}/conf/${selection} ]; then
+        peer_desc=$(eval echo $(jq '.description' ${config_folder}/conf/${selection}))
+        peer_key=$(eval echo $(jq '.public_key' ${config_folder}/conf/${selection}))
+        peer_host=$(eval echo $(jq '.endpoint_host' ${config_folder}/conf/${selection}))
+        uci set network.peer_surfshark.description=${peer_desc}
+        uci set network.peer_surfshark.public_key=${peer_key}
+        uci set network.peer_surfshark.endpoint_host=${peer_host}
+        uci commit network
+        /etc/init.d/network restart
+        cp -f "${config_folder}/conf/${selection}" ${config_folder}/surfshark
     else
-        echo "No old keys or profiles found."
+        echo "server conf not recognised"
+        surfshark_switch
     fi
 }
 
-read_config1
+reset_surfshark() {
+    # set everything back to a blank state excluding config.json
+    echo "removing network settings"
+    uci -q delete network.sswg
+    uci -q delete network.surfshark
+    uci -q delete network.wgs_surfshark
+    uci -q delete network.peer_surfshark
+    uci commit network
+    /etc/init.d/network restart
+    echo "removing firewall settings"
+    uci -q delete firewall.sswg_zone
+    uci -q delete firewall.sswg_forwarding
+    uci commit firewall
+    /etc/init.d/firewall restart
+    echo "removing generated configuration files"
+    rm -fr ${config_folder}/conf
+    rm -f ${config_folder}/*servers.json
+    rm -f ${config_folder}/wg.json
+    rm -f ${config_folder}/token.json
+    rm -f ${config_folder}/token_expires.json
+    rm -f ${config_folder}/surfshark
+}
+
+read_config
 parse_arg "$@"
 
 if [ $reset_all -eq 1 ]; then
@@ -357,7 +356,6 @@ if [ $reset_all -eq 1 ]; then
 fi
 
 if [ $generate_servers -eq 1 ]; then
-    read_config2
     get_servers
     gen_client_confs
     echo "server list now:"
@@ -365,18 +363,16 @@ if [ $generate_servers -eq 1 ]; then
     exit 1
 fi
 
-#if [ $switch_conf -eq 1 ]; then
-#    surfshark_up
-#    exit 1
-#fi
+if [ $switch_conf -eq 1 ]; then
+    surfshark_switch
+    exit 1
+fi
 
 echo "Logging in if needed ..."
 wg_login
 
 echo "Generating keys ..."
 wg_gen_keys
-
-read_config2
 
 if [ $register -eq 1 ]; then
     echo "Registring pubkey ..."
@@ -399,7 +395,7 @@ echo "Generating profiles ..."
 fi
 
 if [ ! -e ${config_folder}/surfshark ]; then
-	surfshark_up
+	surfshark_1st_up
 fi
 
 echo "Done!"
